refactor: put the mocks under certificate-orchestrator
