+++ crates/agglayer-aggregator-notifier/Cargo.toml
@@ -37,6 +37,7 @@ sp1-sdk.workspace = true
 sp1-prover.workspace = true
 
 [dev-dependencies]
+agglayer-certificate-orchestrator = { path = "../agglayer-certificate-orchestrator", features = ["testutils"] }
 agglayer-storage = { path = "../agglayer-storage", features = ["testutils"] }
 pessimistic-proof = { workspace = true, features = ["testutils"] }
 pessimistic-proof-test-suite = { path = "../pessimistic-proof-test-suite" }
@@ -55,4 +56,4 @@ url.workspace = true
 [features]
 coverage = []
 sp1 = []
-testutils = []
+testutils = ["agglayer-certificate-orchestrator/testutils"]
